+++ internal/engine/netx/resolver/integration_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/apex/log"
 	"github.com/ooni/probe-cli/v3/internal/engine/netx/resolver"
+	"github.com/ooni/probe-cli/v3/internal/netxlite"
 )
 
 func init() {
@@ -18,7 +19,7 @@ func testresolverquick(t *testing.T, reso resolver.Resolver) {
 	if testing.Short() {
 		t.Skip("skip test in short mode")
 	}
-	reso = resolver.LoggingResolver{Logger: log.Log, Resolver: reso}
+	reso = netxlite.ResolverLogger{Logger: log.Log, Resolver: reso}
 	addrs, err := reso.LookupHost(context.Background(), "dns.google.com")
 	if err != nil {
 		t.Fatal(err)
@@ -44,7 +45,7 @@ func testresolverquickidna(t *testing.T, reso resolver.Resolver) {
 		t.Skip("skip test in short mode")
 	}
 	reso = resolver.IDNAResolver{
-		resolver.LoggingResolver{Logger: log.Log, Resolver: reso},
+		netxlite.ResolverLogger{Logger: log.Log, Resolver: reso},
 	}
 	addrs, err := reso.LookupHost(context.Background(), "яндекс.рф")
 	if err != nil {
@@ -56,7 +57,7 @@ func testresolverquickidna(t *testing.T, reso resolver.Resolver) {
 }
 
 func TestNewResolverSystem(t *testing.T) {
-	reso := resolver.SystemResolver{}
+	reso := netxlite.ResolverSystem{}
 	testresolverquick(t, reso)
 	testresolverquickidna(t, reso)
 }
